+++ apps/expo/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "movie-web",
+  "name": "@movie-web/mobile",
   "version": "0.1.0",
   "private": true,
   "main": "expo-router/entry",
