+++ packages/tmdb/src/index.ts
@@ -1,2 +1,4 @@
 export const name = "tmdb";
 export * from "./search";
+export * from "./details";
+export * from "./util";
